+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support `str/bytes.startswith/endswith(tuple)`
   - Support `str/bytes.maketrans/translate`
   - Support `glob.escape`
+  - Support `base64.{b16encode, b16decode}`
 
 ### Fixed